+++ src/Http/Controllers/HistoryActionController.php
@@ -3,10 +3,15 @@
 namespace D3Creative\Sentinel\Http\Controllers;
 
 use D3Creative\Sentinel\Services\HistoryService;
+use D3Creative\Sentinel\Http\Controllers\Concerns\GuardsSentinelActions;
 use Statamic\Http\Controllers\CP\ActionController;
 
 class HistoryActionController extends ActionController
 {
+    use GuardsSentinelActions;
+
+    const SENTINEL_ACTION = 'delete_history_entry';
+
     protected function getSelectedItems($items, $context)
     {
         $all = collect(app(HistoryService::class)->all());